Pre-fill the correct input when accepting a revision.
Fix #6

When accepting the revision, we were doing a simple query
to retrieve the project id input and prefill it with the
testing tag prefix.
But it might happen that the user have other inputs displayed (e.g. Reviewers),
and the selector we're using is the same, so we'd fill the wrong input.

Unfortunately, there's no meaningful attributes that could only target the
project tags input.
Luckily, there's a <data> element on the page that has some data about project
ids, and more specifically, the markup that will be used to render the input,
where what seems to be a unique data-meta id is used.
So here, at startup we retrieve this id, and when accepting the revision, we
check the element with this data-meta attribute, and retrieve the sibling
input.